add: tost 적용
